+++ lib/contact_card.dart
@@ -1,4 +1,5 @@
 library contact_card;
 
 export 'cards/mecard.dart';
+export 'cards/vcard.dart';
 export 'errors/errors.dart';